Move cnl::from_value into a separate header file
