+++ include/cnl/num_traits.h
@@ -14,6 +14,7 @@
 #include "limits.h"
 
 #include "bits/num_traits/digits.h"
+#include "bits/num_traits/from_value.h"
 #include "bits/num_traits/set_digits.h"
 #include "bits/power.h"
 #include "bits/type_traits.h"
@@ -134,53 +135,6 @@ namespace cnl {
             return for_rep<Result>(f, to_rep(std::forward<Args>(args))...);
         }
     }
-
-    ////////////////////////////////////////////////////////////////////////////////
-    // cnl::from_value
-
-    namespace _impl {
-        template<typename Result, typename Value>
-        struct from_value_simple {
-            constexpr Result operator()(Value const& value) const {
-                return value;
-            }
-        };
-    }
-
-    template<typename Number, typename Value, class Enable = void>
-    struct from_value : _impl::from_value_simple<void, Value> {
-        void operator()(Value const &) const;
-    };
-
-    template<class Number, class Value>
-    struct from_value<
-            Number, Value, _impl::enable_if_t<cnl::is_integral<Number>::value && cnl::is_integral<Value>::value>>
-            : _impl::from_value_simple<Value, Value> {
-    };
-
-    template<class Number, CNL_IMPL_CONSTANT_VALUE_TYPE Value>
-    struct from_value<Number, constant<Value>, _impl::enable_if_t<is_integral<Number>::value>> {
-    private:
-        using _result_type = set_digits_t<
-                make_signed_t<Number>,
-                _impl::max(digits<int>::value, _impl::used_digits(Value))>;
-    public:
-        constexpr _result_type operator()(constant<Value> const &value) const {
-            return _result_type(value);
-        }
-    };
-
-    namespace _impl {
-        template<typename Number, typename Value>
-        constexpr auto make_number(Value const& value)
-        -> decltype(cnl::from_value<Number, Value>{}(value))
-        {
-            return cnl::from_value<Number, Value>{}(value);
-        }
-    }
-
-    template<typename Number, typename Value>
-    using from_value_t = decltype(_impl::make_number<Number>(std::declval<Value>()));
 
     ////////////////////////////////////////////////////////////////////////////////
     // cnl::shift
